Resolvers for the dynamic json schema for each type in the type system + Resolver for the schema identifier + Refactored field resolver of the entity types and object resolvers of the flow types
- GraphQL: Added resolver for the dynamic JSON schema for each type in the type system
- GraphQL: Added resolver for the dynamic JSON schema identifier ($id) for each type in the type system
- Dynamic Graph: Refactored field resolvers of the entity types
- Dynamic Graph: Refactored object resolvers of the flow types
