+++ crates/dynamic-graph/impl/src/field/entity/inbound/mod.rs
@@ -1,0 +1,3 @@
+pub mod outbound_entities_field;
+pub mod outbound_field;
+pub mod relation_field;
