+++ crates/dynamic-graph/impl/src/object/entity/query/mod.rs
@@ -1,13 +1,13 @@
 use crate::extension::divergent::is_divergent;
 use crate::extension::field_description::get_dynamic_graph_field_descriptions;
 use crate::extension::field_name::get_dynamic_graph_field_names;
-use crate::field::entity_id_field;
-use crate::field::entity_inbound_relation_field;
-use crate::field::entity_outbound_relation_field;
-use crate::field::entity_property_field;
-use crate::field::inbound_entity_to_outbound_field;
+use crate::field::id::entity_id_field;
+use crate::field::inbound::outbound_field::inbound_entity_to_outbound_field;
+use crate::field::inbound::relation_field::entity_inbound_relation_field;
 use crate::field::instance_component_id_field;
-use crate::field::outbound_entity_to_inbound_field;
+use crate::field::outbound::inbound_field::outbound_entity_to_inbound_field;
+use crate::field::outbound::relation_field::entity_outbound_relation_field;
+use crate::field::property::property_field::entity_property_field;
 use crate::interface::entity::INTERFACE_ENTITY;
 use crate::object::types::DynamicGraphTypeDefinition;
 use async_graphql::dynamic::Object;
