+++ crates/dynamic-graph/impl/src/object/flow/mod.rs
@@ -1,0 +1,2 @@
+pub mod mutation;
+pub mod query;
